+++ src/birdnet_tests/helper.py
@@ -10,6 +10,8 @@ from ordered_set import OrderedSet
 from birdnet.types import Species, SpeciesPrediction, SpeciesPredictions, TimeInterval
 
 TEST_FILES_DIR = Path("src/birdnet_tests/test_files")
+# Duration: 120s
+TEST_FILE_WAV = TEST_FILES_DIR / "soundscape.wav"
 
 
 def species_prediction_is_equal(