+++ src/birdnet_tests/models/model_v2m4_tflite_custom_py/test_custom_model_v2m4_tflite.py
@@ -4,10 +4,7 @@ import numpy.testing as npt
 import pytest
 
 from birdnet.models.model_v2m4_tflite_custom import CustomModelV2M4TFLite
-from birdnet_tests.helper import TEST_FILES_DIR
-
-# Duration: 120s
-TEST_FILE_WAV = TEST_FILES_DIR / "soundscape.wav"
+from birdnet_tests.helper import TEST_FILE_WAV
 
 
 def test_invalid_classifier_name_raises_value_error():
@@ -42,21 +39,37 @@ def test_minimum_test_soundscape_predictions_are_correct():
   assert list(res[(0, 3)].keys())[0] == 'Poecile atricapillus_Black-capped Chickadee'
   npt.assert_almost_equal(
     res[(0, 3)]['Poecile atricapillus_Black-capped Chickadee'],
-    0.7648012,
+    0.8326453,
     decimal=6
   )
 
   assert list(res[(66, 69)].keys())[0] == 'Junco hyemalis_Dark-eyed Junco'
   npt.assert_almost_equal(
     res[(66, 69)]['Junco hyemalis_Dark-eyed Junco'],
-    0.20300059,
+    0.19126873,
     decimal=6
   )
 
-  assert list(res[(117, 120)].keys())[0] == 'Cyanocitta cristata_Blue Jay'
+  assert list(res[(117, 120)].keys())[0] == 'Junco hyemalis_Dark-eyed Junco'
   npt.assert_almost_equal(
-      res[(117, 120)]['Cyanocitta cristata_Blue Jay'],
-      0.14364028,
+      res[(117, 120)]['Junco hyemalis_Dark-eyed Junco'],
+      0.14393075,
       decimal=6
     )
+  assert len(res) == 40
+
+
+def test_no_sigmoid_soundscape_predictions_are_correct():
+  classifier_folder = Path("src/birdnet_tests/test_files/custom_model_v2m4_tflite")
+  model = CustomModelV2M4TFLite(classifier_folder, "CustomClassifier")
+
+  res = model.predict_species_within_audio_file(
+    TEST_FILE_WAV, min_confidence=0, apply_sigmoid=False)
+
+  npt.assert_almost_equal(
+    res[(0, 3)]['Poecile atricapillus_Black-capped Chickadee'],
+    1.6044917,
+    decimal=6
+  )
+
   assert len(res) == 40