+++ src/birdnet_tests/models/test_predict_species_within_audio_file.py
@@ -13,11 +13,10 @@ from tqdm import tqdm
 from birdnet.models.model_v2m4_base import ModelV2M4Base
 from birdnet.models.model_v2m4_protobuf import ModelV2M4Protobuf
 from birdnet.types import Language, Species, SpeciesPredictions
-from birdnet_tests.helper import (TEST_FILES_DIR, convert_predictions_to_numpy,
+from birdnet_tests.helper import (TEST_FILE_WAV, TEST_FILES_DIR, convert_predictions_to_numpy,
                                   species_predictions_are_equal)
 
 # Duration: 120s
-TEST_FILE_WAV = TEST_FILES_DIR / "soundscape.wav"
 TEST_PATH = Path(f"{TEST_FILE_WAV}.global.pkl")